Move OsmJsonReader to proper place
