+++ hoot-core/hoot-core.pro
@@ -497,7 +497,7 @@ SOURCES += \
     src/main/cpp/hoot/core/io/ElementComparer.cpp \
     src/main/cpp/hoot/core/cmd/ApplyChangesetCmd.cpp \
     src/main/cpp/hoot/core/io/OsmApiDbChangesetWriter.cpp \
-    src/main/cpp/hoot/core/util/OsmJsonReader.cpp
+    src/main/cpp/hoot/core/io/OsmJsonReader.cpp
 
 HEADERS += \
     src/main/cpp/hoot/core/util/Progress.h \
@@ -995,5 +995,5 @@ HEADERS += \
     src/main/cpp/hoot/core/io/OsmChangesetSqlFileWriter.h \
     src/main/cpp/hoot/core/ops/FindIntersectionsOp.h \
     src/main/cpp/hoot/core/io/OsmApiDbChangesetWriter.h \
-    src/main/cpp/hoot/core/util/OsmJsonReader.h
+    src/main/cpp/hoot/core/io/OsmJsonReader.h
