+++ hoot-core/src/main/cpp/hoot/core/io/OsmJsonReader.cpp
