+++ hoot-core/src/main/cpp/hoot/core/io/OsmJsonReader.h
